renamed editLoadControler to controllers/editor
