+++ meteor/routes/routes.js
@@ -6,21 +6,21 @@
 // global route settings
 Router.configure({
     // Template displayed while loading data.
-    loadingTemplate: 'loading',
-    // Template displayed when there's no route for the sub domain.
-    notFoundTemplate: 'notFound',
+    loadingTemplate: "loading",
+    // Template displayed when there"s no route for the sub domain.
+    notFoundTemplate: "notFound",
 });
 
 // route settings for default endpoint "/"
-Router.route('/', {
-    name: 'editor',
-    template: 'alloyEditor',
-    where: 'client',
+Router.route("/", {
+    name: "editor",
+    template: "alloyEditor",
+    where: "client",
 });
 
 // route settings for endpoint with model id "/:_id"
-Router.route('/:_id', {
-    name: 'editorLoad',
-    controller: "editorLoadController",
-    where: 'client'
+Router.route("/:_id", {
+    name: "editorLoad",
+    controller: "editor",
+    where: "client"
 });
